adding slider to change window size

+++ Application/src/main/java/Graphing/Graph_ECG.java
@@ -6,7 +6,7 @@ import simulation.Value_Counter;
 public class Graph_ECG extends Graph {
     private ECG ecgdata;
 
-    public Graph_ECG(String colorGraph, Value_Counter obj, double sample_period, float time_shown, ECG ecgdata_input){
+    public Graph_ECG(String colorGraph, Value_Counter obj, double sample_period, int time_shown, ECG ecgdata_input){
         super(colorGraph, obj, sample_period, time_shown);
         ecgdata = ecgdata_input;
     }
