+++ Application/src/main/java/Graphing/Graph_Pressure.java
@@ -6,7 +6,7 @@ import simulation.Value_Counter;
 public class Graph_Pressure extends Graph {
     private Blood_Pressure pressuredata;
 
-    public Graph_Pressure(String colorGraph, Value_Counter obj, double sample_period, float time_shown, Blood_Pressure pressuredata_input){
+    public Graph_Pressure(String colorGraph, Value_Counter obj, double sample_period, int time_shown, Blood_Pressure pressuredata_input){
         super(colorGraph, obj, sample_period, time_shown);
         pressuredata = pressuredata_input;
     }
